[100%] :) Modified the solver to use the correct critical optimization
The Board changes have been reverted as in this program we are using the iterator only once
In the solver the previous approach of saving the last two boards did not work as we are dealing with priority queue and we don't exactly know what board pops out of the queue, so our last2last board will be wrong.
The manhattan distance gives better representation of the heuristic than hamming distance. It greatly improves performance.